pyproject.toml
[tool.poetry]
name = "pysimdjson-schemaful"
version = "0.2.1"
description = "Schema-aware pysimdjson loader for efficient parsing of large excessive inputs."
authors = ["Tzoiker <tzoiker@gmail.com>"]
repository = "https://github.com/tzoiker/pysimdjson-schemaful"
license = "MIT"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python",
    "Topic :: Internet",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Typing :: Typed",
]
readme = "README.md"
packages = [
  {include = "simdjson_schemaful"}
]

[tool.commitizen]
version_provider = "poetry"
version_scheme = "pep440"
tag_format = "v$version"
changelog_incremental = true
version_files = [
    "simdjson_schemaful/__version__.py",
]


[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pysimdjson = [
    {version=">=2,<6", python=">=3.8,<3.9"},
    {version=">=3,<6", python=">=3.9,<3.11"},
    # version 4 does not build for 3.11 for some reason
    {version=">=3,!=4.*,<6", python=">=3.11,<3.12"},
]
pydantic = { version = ">=1,<3", optional = true }

[tool.poetry.extras]
pydantic = ["pydantic"]

[tool.poetry.group.dev.dependencies]
attrs = "^23.1.0"
black = "^23.10.0"
commitizen = "^3.12.0"
markdown-pytest = "^0.3.0"
mypy = "^1.0"
orjson = "^3.9.9"
pre-commit = "^3.3.1"
pydantic = "^1.0"
pytest = "^7.4.2"
pytest-cov = "^4.1"
ruff = "^0.0.291"
tox = "^4.0"

[tool.mypy]
plugins = "pydantic.mypy"
allow_subclassing_any = true
allow_untyped_calls = true
allow_untyped_decorators = true
ignore_missing_imports = true
implicit_reexport = true
pretty = true
show_error_codes = true
strict = true
warn_return_any = false
files = [
    "simdjson_schemaful",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.ini_options]
python_files = "test_*"
python_functions = "test_*"
python_classes = "Test*"
addopts = "-p no:cacheprovider"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["simdjson_schemaful"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
  "BLE",
  "C90",
  "E",
  "F",
  "G",
  "I",
  "ICN",
  "ISC",
  "PLE",
  "Q",
  "RUF006",
  "RUF100",
  "T10",
  "T20",
  "TID",
  "UP",
  "W",
]
